Dokument - Implementation cleanup

+++ src/ReceptiAPI/DTO/NamirnicaDTO.cs
@@ -13,16 +13,16 @@ namespace ReceptiAPI.DTO
         [JsonProperty(PropertyName = "kategorija")]
         public string Kategorija { get; set; }
         [JsonProperty(PropertyName = "kalorije")]
-        public ulong Kalorije { get; set; }
+        public decimal Kalorije { get; set; }
         [JsonProperty(PropertyName = "proteini")]
-        public ulong Proteini { get; set; }
+        public decimal Proteini { get; set; }
         [JsonProperty(PropertyName = "masti")]
-        public ulong Masti { get; set; }
+        public decimal Masti { get; set; }
         [JsonProperty(PropertyName = "zasiceneMasti")]
-        public ulong ZasiceneMasti { get; set; }
+        public decimal ZasiceneMasti { get; set; }
         [JsonProperty(PropertyName = "seceri")]
-        public ulong Seceri { get; set; }
+        public decimal Seceri { get; set; }
         [JsonProperty(PropertyName = "vlakna")]
-        public ulong Vlakna { get; set; }
+        public decimal Vlakna { get; set; }
     }
 }
